refactor(identity): restructure module to standard 4-layer layout
Move services.py into domain/, extract LoginRequest/TokenResponse into
api/schemas.py, add empty infrastructure/ and ports/ layers. Old
services.py kept as re-export shim for backward compatibility.
