+++ backend/src/hiresense/identity/api/dependencies.py
@@ -5,7 +5,7 @@ from typing import Annotated
 from fastapi import Depends, HTTPException, status
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 
-from hiresense.identity.services import AuthService
+from hiresense.identity.domain import AuthService
 
 security = HTTPBearer()
