+++ backend/src/hiresense/identity/infrastructure/__init__.py
@@ -1,0 +1,1 @@
+"""Identity infrastructure layer."""
