+++ backend/src/hiresense/identity/ports/__init__.py
@@ -1,0 +1,1 @@
+"""Identity ports layer."""
